Fix #70 by bypassing lazy loading of datasets (#71)
By default, rioxarray lazyily loads datasets. On Windows, this causes
a PermissionError when the temporary directory containing the downloaded
files is deleted, as it is still open in rioxarray. To avoid that issue,
we now explicitly load the dataset into memory, allowing tempfiles to be
removed. This will potentially increase memory usage in some use cases, but
given the size limitations of datasets downloaded from Earth Engine, that
should never be a practical problem.