Remove deprecated `path` argument (#88)
This was a shortcut for writing directly to NetCDF that's been
deprecated for almost two years. Since removing it will shift the
order of positional arguments, I went ahead and made the arguments
keyword-only at the same time. I suspect that's how the vast
majority of users were calling it anyways, since the first argument
was the deprecated "path".